Move add link form to dedicated page
