+++ assets/style.css
@@ -51,8 +51,11 @@ textarea {
   }
 .content {background:#fff;color:#000;padding:20px;}
 
-.open-modal{background:#1DA1F2;color:#fff;border:none;width:50px;height:50px;display:flex;align-items:center;justify-content:center;cursor:pointer;font-size:36px;margin:0 5px;flex-shrink:0;padding:0;}
-.open-modal svg{width:24px;height:24px;}
+.open-modal,
+.add-link-btn{background:#1DA1F2;color:#fff;border:none;width:50px;height:50px;display:flex;align-items:center;justify-content:center;cursor:pointer;font-size:36px;margin:0 5px;flex-shrink:0;padding:0;text-decoration:none;}
+.open-modal svg,
+.add-link-btn svg{width:24px;height:24px;}
+.add-link-btn{border-radius:50%;}
 .add-mobile{display:none;}
 @media (max-width:600px){.add-mobile{display:flex;}.top-menu .menu .add-menu{display:none;}}
 .search-toggle{background:#fff;color:#1DA1F2;border:none;border-radius:4px;width:32px;height:32px;display:flex;align-items:center;justify-content:center;cursor:pointer;font-size:20px;margin-left:5px;flex-shrink:0;}
@@ -64,7 +67,12 @@ textarea {
 .add-modal.show{display:flex;}
 .add-modal-content{background:#fff;color:#000;padding:10px;padding-left:20px;border-radius:20px;max-width:500px;width:100%;position:relative;max-height:100%;overflow:auto;}
 .modal-close{position:absolute;top:10px;right:10px;background:none;border:none;font-size:24px;cursor:pointer;}
-.add-modal-content h2.modal-title{text-align:center;margin:0 0 20px;color:#1DA1F2;}
+.add-modal-content h2.modal-title,
+.favolink-card .modal-title{text-align:center;margin:0 0 20px;color:#1DA1F2;}
+.favolink-page{min-height:calc(100vh - 80px);display:flex;align-items:center;justify-content:center;padding:40px 20px;background:#f5f8fa;}
+.favolink-card{background:#fff;color:#000;border-radius:20px;max-width:480px;width:100%;padding:35px 30px;box-shadow:0 10px 25px rgba(0,0,0,0.1);}
+.favolink-tagline{text-align:center;margin:0 0 10px;color:#1DA1F2;font-weight:700;}
+.favolink-card .control-forms{margin-top:10px;}
 .control-forms{display:flex;flex-direction:column;gap:10px;}
 .control-forms .form-section{display:flex;flex-direction:column;gap:10px;align-items:center;}
 .control-forms .form-section h3{text-align:center;margin:0;color:#6c757d;}
