+++ header.php
@@ -26,14 +26,14 @@ $jsVersion  = filemtime(__DIR__ . '/assets/main.js');
     <div class="logo"><a href="/panel.php"><img src="/img/linkaloo_white.png" alt="linkaloo"><!-- Logo file already on server --></a></div>
     <nav>
         <?php if(isset($_SESSION['user_id']) && isset($categorias)): ?>
-            <button type="button" class="open-modal add-mobile" aria-label="Añadir"><i data-feather="plus"></i></button>
+            <a href="/agregar_favolink.php" class="add-link-btn add-mobile" aria-label="Añadir"><i data-feather="plus"></i></a>
         <?php endif; ?>
         <button class="menu-toggle" aria-label="Menú"><span></span><span></span><span></span></button>
         <ul class="menu">
             <?php if(isset($_SESSION['user_id'])): ?>
                 <?php if(isset($categorias)): ?>
                     <li class="add-menu">
-                        <button type="button" class="open-modal" aria-label="Añadir"><i data-feather="plus"></i></button>
+                        <a href="/agregar_favolink.php" class="add-link-btn" aria-label="Añadir"><i data-feather="plus"></i></a>
                     </li>
                 <?php endif; ?>
                 <li><a href="/tableros.php">Tableros</a></li>
@@ -55,31 +55,4 @@ $jsVersion  = filemtime(__DIR__ . '/assets/main.js');
         </ul>
     </nav>
 </header>
-<?php if(isset($categorias)): ?>
-<div class="add-modal">
-    <div class="add-modal-content" style="padding: 30px;">
-        <button type="button" class="modal-close" aria-label="Cerrar">&times;</button>
-        <div class="app-logo"><img src="/img/logo_linkaloo_blue.png" alt="Linkaloo logo"></div>
-        <h2 class="modal-title">Añadir tu favolink</h2>
-        <div class="control-forms">
-            <div class="form-section">
-                <form method="post" class="form-link">
-                    <input type="url" name="link_url" placeholder="Pega aquí el link" required>
-                    <input type="text" name="link_title" placeholder="Título (opcional)" maxlength="50">
-                    <div class="select-create">
-                        <select name="categoria_id">
-                            <option value="">Elige el tablero</option>
-                            <?php foreach($categorias as $categoria): ?>
-                                <option value="<?= $categoria['id'] ?>"><?= htmlspecialchars($categoria['nombre']) ?></option>
-                            <?php endforeach; ?>
-                        </select>
-                        <input type="text" name="categoria_nombre" placeholder="o crea un nuevo (opcional)">
-                    </div>
-                    <button type="submit">Guardar favolink</button>
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-<?php endif; ?>
 <div class="content">
