+++ panel.php
@@ -1,7 +1,6 @@
 <?php
 require 'config.php';
 require 'favicon_utils.php';
-require_once 'image_utils.php';
 require_once 'session.php';
 require_once 'device.php';
 if(!isset($_SESSION['user_id'])){
@@ -18,131 +17,14 @@ $descLimit = isMobile() ? 50 : 150;
 $error = $_SESSION['panel_error'] ?? '';
 unset($_SESSION['panel_error']);
 
-function ensureUtf8($string){
-    $encoding = mb_detect_encoding($string, 'UTF-8, ISO-8859-1, WINDOWS-1252', true);
-    if($encoding && $encoding !== 'UTF-8'){
-        $string = mb_convert_encoding($string, 'UTF-8', $encoding);
+if(isset($_GET['shared'])){
+    $sharedIncoming = trim($_GET['shared']);
+    if(isValidSharedUrl($sharedIncoming)){
+        $redirect = 'agregar_favolink.php?shared=' . rawurlencode($sharedIncoming);
+        header('Location: ' . $redirect);
+        exit;
     }
-    return $string;
 }
-
-function canonicalizeUrl($url){
-    $parts = parse_url(trim($url));
-    if(!$parts || empty($parts['host'])){
-        return $url;
-    }
-    $scheme = strtolower($parts['scheme'] ?? 'http');
-    $host = strtolower($parts['host']);
-    $path = isset($parts['path']) ? rtrim($parts['path'], '/') : '';
-    $query = isset($parts['query']) ? '?' . $parts['query'] : '';
-    $port = isset($parts['port']) ? ':' . $parts['port'] : '';
-    return $scheme . '://' . $host . $port . $path . $query;
-}
-
-function scrapeMetadata($url){
-    $ch = curl_init($url);
-    curl_setopt_array($ch, [
-        CURLOPT_RETURNTRANSFER => true,
-        CURLOPT_FOLLOWLOCATION => true,
-        CURLOPT_USERAGENT => 'Mozilla/5.0 (compatible; linkalooBot/1.0)',
-        CURLOPT_TIMEOUT => 5,
-    ]);
-    $html = curl_exec($ch);
-    curl_close($ch);
-    if(!$html){
-        return [];
-    }
-    $enc = mb_detect_encoding($html, 'UTF-8, ISO-8859-1, WINDOWS-1252', true);
-    if($enc){
-        $html = mb_convert_encoding($html, 'HTML-ENTITIES', $enc);
-    }
-    libxml_use_internal_errors(true);
-    $doc = new DOMDocument();
-    $doc->loadHTML($html);
-    $xpath = new DOMXPath($doc);
-    $getMeta = function($name, $attr='property') use ($xpath){
-        $nodes = $xpath->query("//meta[@$attr='$name']/@content");
-        return $nodes->length ? trim($nodes->item(0)->nodeValue) : '';
-    };
-    $meta = [];
-    $titles = $doc->getElementsByTagName('title');
-    if($titles->length){
-        $meta['title'] = trim($titles->item(0)->textContent);
-    }
-    $meta['description'] = $getMeta('og:description') ?: $getMeta('description','name');
-    $meta['image'] = $getMeta('og:image') ?: $getMeta('twitter:image');
-    if(!empty($meta['image']) && !preg_match('#^https?://#', $meta['image'])){
-        $parts = parse_url($url);
-        $base = $parts['scheme'].'://'.$parts['host'];
-        if(isset($parts['port'])){
-            $base .= ':'.$parts['port'];
-        }
-        $meta['image'] = rtrim($base,'/').'/'.ltrim($meta['image'],'/');
-    }
-    foreach($meta as &$value){
-        $value = ensureUtf8($value);
-    }
-    unset($value);
-    return $meta;
-}
-
-if($_SERVER['REQUEST_METHOD'] === 'POST'){
-    if(isset($_POST['link_url'])){
-        $link_url = trim($_POST['link_url']);
-        $link_title = trim($_POST['link_title']);
-        $categoria_id = isset($_POST['categoria_id']) ? (int)$_POST['categoria_id'] : 0;
-        $categoria_nombre = trim($_POST['categoria_nombre'] ?? '');
-        if($categoria_nombre){
-            $stmt = $pdo->prepare('INSERT INTO categorias (usuario_id, nombre) VALUES (?, ?)');
-            $stmt->execute([$user_id, $categoria_nombre]);
-            $categoria_id = (int)$pdo->lastInsertId();
-        }
-        if($link_url && $categoria_id){
-            $meta = scrapeMetadata($link_url);
-            if(!$link_title && !empty($meta['title'])){
-                $link_title = $meta['title'];
-            }
-            $link_title = ensureUtf8($link_title);
-            if (mb_strlen($link_title) > 50) {
-                $link_title = mb_substr($link_title, 0, 47) . '...';
-            }
-            $descripcion = ensureUtf8($meta['description'] ?? '');
-            if (mb_strlen($descripcion) > $descLimit) {
-                $descripcion = mb_substr($descripcion, 0, $descLimit - 3) . '...';
-            }
-            $imagen = $meta['image'] ?? '';
-            if (empty($imagen)) {
-                $domain = parse_url($link_url, PHP_URL_HOST);
-                if ($domain) {
-                    $imagen = getLocalFavicon($domain);
-                }
-            }
-            if(!empty($imagen) && str_starts_with($imagen, 'http')){
-                $localImg = saveImageFromUrl($imagen, $user_id);
-                if($localImg){
-                    $imagen = $localImg;
-                }
-            }
-            $canon = canonicalizeUrl($link_url);
-            $hash = sha1($canon);
-            $check = $pdo->prepare('SELECT id FROM links WHERE usuario_id = ? AND hash_url = ?');
-            $check->execute([$user_id, $hash]);
-            if($check->fetch()){
-                $_SESSION['panel_error'] = 'Este link ya está guardado.';
-            } else {
-                $stmt = $pdo->prepare('INSERT INTO links (usuario_id, categoria_id, url, url_canonica, titulo, descripcion, imagen, hash_url) VALUES (?, ?, ?, ?, ?, ?, ?, ?)');
-                $stmt->execute([$user_id, $categoria_id, $link_url, $canon, $link_title, $descripcion, $imagen, $hash]);
-                if ($stmt->rowCount()) {
-                    $upd = $pdo->prepare('UPDATE categorias SET modificado_en = NOW() WHERE id = ?');
-                    $upd->execute([$categoria_id]);
-                }
-            }
-        }
-    }
-    header('Location: panel.php');
-    exit;
-}
-
 $stmt = $pdo->prepare('SELECT id, nombre FROM categorias WHERE usuario_id = ? ORDER BY modificado_en DESC');
 $stmt->execute([$user_id]);
 $categorias = $stmt->fetchAll();
